--- Sesion-03/Postwork/Readme.md
@@ -1,48 +1,0 @@
-# Postwork 03
-
-## 🎩 Objetivo
-
-- Practicar lo aprendido sobre programación funcional.
-- Hacer uso de la programación funcional para generar y ordenar una lista de datos (descendente y ascendente).
-
-## 🎯 Requisitos
-
-- MySQL instalado y configurado
-- Apache Maven 3.8.4 o superior
-- JDK (o OpenJDK)
-
-## 🚀 Desarrollo
-
-**Realizar en equipo**
-
-Completa las siguientes instrucciones para completar el tercer postwork:
-
-1. Utilicen el proyecto del postwork 2.
-
-2. Creen un nuevo package con el nombre de tools.
-
-3. Un miembro del equipo deberá generar la clase **ReporteCalificaciones** la cual se encargará de:
-
-    - Generar la lista
-    - Ordenar alfabéticamente (Curso)
-    - Ordenar por calificaciones (Descendente)
-    - ***Recuerda utilizar map, sort, collect**
-
-4. Otra persona realizará la prueba **ReporteCalificacionesTest** la cual se encargará de generar:
-
-    - Materia
-    - Estudiantes (Por lo menos tres)
-    - Curso
-    - Enviar la información a la clase creada en el paso tres
-
-5. Por último, el último miembro realizará las siguientes pruebas a **ReporteCalificacionesTest**:
-
-    - Ordena alfabéticamente (Curso)
-    - Ordena por calificaciones (Descendente)
-
-<br/>
-<br/>
-
-[Regresar ](../Readme.md)(Sesión 03)
-
-[Siguiente ](../../Sesion-04/Readme.md)(Sesión 04)
