--- Sesion-03/Postwork/base/src/main/java/org/bedu/postwork/javase2project/persistence/EstudianteRepository.java
@@ -1,9 +1,0 @@
-package org.bedu.postwork.javase2project.persistence;
-
-import org.bedu.postwork.javase2project.model.Curso;
-import org.bedu.postwork.javase2project.model.Estudiante;
-import org.springframework.data.repository.CrudRepository;
-
-public interface EstudianteRepository extends CrudRepository<Estudiante, Long> {
-    Estudiante findById(long id);
-}
