--- Sesion-03/Postwork/base/src/main/java/org/bedu/postwork/javase2project/persistence/MateriaRepository.java
@@ -1,9 +1,0 @@
-package org.bedu.postwork.javase2project.persistence;
-
-import org.bedu.postwork.javase2project.model.Materia;
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface MateriaRepository extends CrudRepository<Materia, Long> {
-}
