--- Sesion-03/Postwork/postwork/src/main/java/org/bedu/postwork/javase2project/Javase2ProjectApplication.java
@@ -1,13 +1,0 @@
-package org.bedu.postwork.javase2project;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class Javase2ProjectApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(Javase2ProjectApplication.class, args);
-	}
-
-}
