--- Sesion-03/Postwork/postwork/src/main/java/org/bedu/postwork/javase2project/model/Estudiante.java
@@ -1,49 +1,0 @@
-package org.bedu.postwork.javase2project.model;
-
-import javax.persistence.*;
-import javax.validation.constraints.Size;
-import java.util.Objects;
-
-@Entity
-@Table(name="estudiantes")
-public class Estudiante {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name="id")
-    private Long id;
-
-    @Column(name="nombre_completo")
-    @Size(max = 45)
-    private String nombreCompleto;
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getNombreCompleto() {
-        return nombreCompleto;
-    }
-
-    public void setNombreCompleto(String nombreCompleto) {
-        this.nombreCompleto = nombreCompleto;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (!(o instanceof Estudiante)) return false;
-        Estudiante that = (Estudiante) o;
-        return id.equals(that.id) &&
-                nombreCompleto.equals(that.nombreCompleto);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(id, nombreCompleto);
-    }
-}
-
