--- Sesion-03/Postwork/postwork/src/main/java/org/bedu/postwork/javase2project/model/Materia.java
@@ -1,48 +1,0 @@
-package org.bedu.postwork.javase2project.model;
-
-import javax.persistence.*;
-import javax.validation.constraints.Size;
-import java.util.Objects;
-
-@Entity
-@Table(name="materias")
-public class Materia {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name="id")
-    private Long id;
-
-    @Column(name = "nombre")
-    @Size(max = 45)
-    private String nombre;
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (!(o instanceof Materia)) return false;
-        Materia materia = (Materia) o;
-        return id.equals(materia.id) &&
-                nombre.equals(materia.nombre);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(id, nombre);
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getNombre() {
-        return nombre;
-    }
-
-    public void setNombre(String nombre) {
-        this.nombre = nombre;
-    }
-}
