--- Sesion-03/Postwork/postwork/src/main/java/org/bedu/postwork/javase2project/multithreading/CalculadorPromedioCurso.java
@@ -1,25 +1,0 @@
-package org.bedu.postwork.javase2project.multithreading;
-
-import org.bedu.postwork.javase2project.model.Curso;
-
-public class CalculadorPromedioCurso implements Runnable {
-
-    private Curso curso;
-    private double promedio;
-
-    public CalculadorPromedioCurso(Curso curso) {
-        this.curso = curso;
-    }
-
-    @Override
-    public void run() {
-        int numAlumnos = 0;
-        for(Integer i : curso.getCalificaciones().values()){
-            promedio += i;
-            numAlumnos++;
-        }
-        promedio /= numAlumnos;
-
-        System.out.println("Promedio del curso: " + curso.getId() + " " + curso.getMateria().getNombre() + " = " + promedio);
-    }
-}
