--- Sesion-03/Postwork/postwork/src/main/java/org/bedu/postwork/javase2project/persistence/EstudianteRepository.java
@@ -1,7 +1,0 @@
-package org.bedu.postwork.javase2project.persistence;
-
-import org.bedu.postwork.javase2project.model.Estudiante;
-import org.springframework.data.repository.CrudRepository;
-
-public interface EstudianteRepository extends CrudRepository<Estudiante, Long> {
-}
