--- Sesion-03/Postwork/postwork/src/main/java/org/bedu/postwork/javase2project/persistence/MateriaRepository.java
@@ -1,7 +1,0 @@
-package org.bedu.postwork.javase2project.persistence;
-
-import org.bedu.postwork.javase2project.model.Materia;
-import org.springframework.data.repository.CrudRepository;
-
-public interface MateriaRepository extends CrudRepository<Materia, Long> {
-}
