--- Sesion-03/Postwork/postwork/src/main/resources/application.properties
@@ -1,3 +1,0 @@
-spring.datasource.url=jdbc:mysql://localhost/javase2_bedu_pw
-spring.datasource.username=pwadmin
-spring.datasource.password=ADMIN_PW_PWD
