--- Sesion-03/Postwork/postwork/src/test/java/org/bedu/postwork/javase2project/Javase2ProjectApplicationTests.java
@@ -1,13 +1,0 @@
-package org.bedu.postwork.javase2project;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class Javase2ProjectApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
