+++ Sesion-03/src/main/java/org/bedu/postwork/javase2project/persistence/EstudianteRepository.java
@@ -1,0 +1,11 @@
+package org.bedu.postwork.javase2project.persistence;
+
+import org.bedu.postwork.javase2project.model.Curso;
+import org.bedu.postwork.javase2project.model.Estudiante;
+import org.springframework.data.repository.CrudRepository;
+
+public interface EstudianteRepository extends CrudRepository<Estudiante, Long> {
+    Estudiante findById(long id);
+    Iterable<Estudiante> findAllByNombreCompleto(String nombreEstudiante);
+
+}
