+++ Sesion-03/src/main/java/org/bedu/postwork/javase2project/persistence/MateriaRepository.java
@@ -1,0 +1,11 @@
+package org.bedu.postwork.javase2project.persistence;
+
+import org.bedu.postwork.javase2project.model.Materia;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface MateriaRepository extends CrudRepository<Materia, Long> {
+    Iterable<Materia> findAllByNombre(String materia);
+
+}
