+++ Sesion-03/src/main/resources/sql/mysql_script.sql
